refactor(postrenderer): move templating logic to HtmlPostRendererTemplate
